replaced float with double

+++ IntroDiagrammes/src/version1/BankingSystem.java
@@ -11,7 +11,7 @@ class BankingSystem {
         account2 = new Account("c2", 200);
     }
 
-    void transfer(float amount, String debitAccountNumber, String creditAcccountNumber) {
+    void transfer(double amount, String debitAccountNumber, String creditAcccountNumber) {
         System.out.println("Nous allons virer un montant de " + amount + " du compte " + debitAccountNumber + " à  " + creditAcccountNumber);
         System.out.println("Etat des comptes avant l'opération:");
         System.out.println(account1.toString());
