+++ IntroDiagrammes/src/version3/BankingSystem.java
@@ -9,7 +9,7 @@ class BankingSystem {
         accountManager = new AccountManager();
     }
 
-    void transfer(float amount, String debitAccountNumber, String creditAcccountNumber) {
+    void transfer(double amount, String debitAccountNumber, String creditAcccountNumber) {
 
         printBefore(amount, debitAccountNumber, creditAcccountNumber);
 
@@ -22,7 +22,7 @@ class BankingSystem {
         printAfter();
     }
 
-    private void printBefore(float amount, String debitAccountNumber, String creditAcccountNumber) {
+    private void printBefore(double amount, String debitAccountNumber, String creditAcccountNumber) {
         System.out.println("Nous allons virer un montant de " + amount + " du compte " + debitAccountNumber + " à  " + creditAcccountNumber);
         System.out.println("Etat des comptes avant l'opération:");
         accountManager.printAccounts();
